+++ price-feeder/CHANGELOG.md
@@ -54,6 +54,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 - [1448](https://github.com/umee-network/umee/pull/1448) Add crypto.com provider.
 - [1496](https://github.com/umee-network/umee/pull/1496) Dynamic provider minimum enforcement with CoinGecko API.
+- [1510](https://github.com/umee-network/umee/pull/1510) Integrate osmosis-api provider into price-feeder.
 
 ### Improvements